sensor, alarm, actu에 필요한 폴더 작업 및 모든 폴더 내부의 타입 통일

+++ backend/src/main/java/com/nova/backend/farm/service/FarmService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface FarmService {
     //2. Farm List NovaId 로 가져오기
-    List<FarmResponseDTO> getFarmListByNovaId(int novaId);
+    List<FarmResponseDTO> getFarmListByNovaId(Long novaId);
 }
